Added streaming logic to TS base class, updated motion correction TS class
